+++ Store/Order.cs
@@ -8,7 +8,7 @@ public class Order
     public int Id => _order.Id;
     public string User => _order.UserName;
     public int TotalCount => Items.Sum(item => item.Count);
-    public float TotalPrice => Items.Sum(item => item.Price * item.Count);
+    public decimal TotalPrice => Items.Sum(item => item.Price * item.Count);
     public OrderItemCollection Items { get; }
 
     public Order(OrderData order)
